ajustements pour linux 10

+++ noethys/Ctrl/CTRL_Ventilation.py
@@ -396,7 +396,7 @@ class RendererCaseMontant(gridlib.GridCellRenderer):
         # Alignement à droite
         largeur, hauteur = dc.GetTextExtent(texte)
         x = rect[0] + rect[2] - largeur - 2
-        y = rect[1] + ((rect[3] - hauteur) / 2.0)
+        y = int(rect[1] + ((rect[3] - hauteur) / 2.0))
         dc.DrawText(texte, x, y)
         
     def GetBestSize(self, grid, attr, dc, row, col):
